# docs/openapi.yaml
openapi: 3.0.2
info:
  title: Chicory API
  version: 1.0.0
  description: API for a multi-tenant platform that enables users to connect data
    sources, build business ontologies, and interact with data through a chat interface.
servers:
- url: https://api.chicory.ai/v1
paths:
  /projects:
    get:
      tags:
        - projects
      summary: List Projects
      description: List all projects, optionally filtered by organization_id
      operationId: list_projects_projects_get
      parameters:
        - name: organization_id
          in: query
          required: true
          schema:
            title: Organization Id
            type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
        - projects
      summary: Create Project
      description: Create a new project
      operationId: create_project_projects_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{project_id}:
    get:
      tags:
        - projects
      summary: Get Project
      description: Get a project by ID
      operationId: get_project_projects__project_id__get
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            title: Project Id
            type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
        - projects
      summary: Update Project
      description: Update a project
      operationId: update_project_projects__project_id__put
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            title: Project Id
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - projects
      summary: Delete Project
      description: Delete a project
      operationId: delete_project_projects__project_id__delete
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            title: Project Id
            type: string
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
...
    ProjectCreate:
      title: ProjectCreate
      description: Schema for creating a project
      type: object
      required:
        - name
        - organization_id
      properties:
        name:
          title: Name
          type: string
          minLength: 1
          maxLength: 100
        organization_id:
          title: Organization Id
          type: string
          minLength: 1
        description:
          title: Description
          type: string
          maxLength: 500
    ProjectList:
      title: ProjectList
      description: Schema for list of projects
      type: object
      required:
        - projects
      properties:
        projects:
          title: Projects
          type: array
          items:
            $ref: '#/components/schemas/ProjectResponse'
    ProjectResponse:
      title: ProjectResponse
      description: Schema for project response
      type: object
      required:
        - id
        - organization_id
        - name
        - created_at
        - updated_at
      properties:
        id:
          title: Id
          type: string
        organization_id:
          title: Organization Id
          type: string
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
        created_at:
          title: Created At
          type: string
          format: date-time
        updated_at:
          title: Updated At
          type: string
          format: date-time
    ProjectUpdate:
      title: ProjectUpdate
      description: Schema for updating a project
      type: object
      properties:
        name:
          title: Name
          type: string
          minLength: 1
          maxLength: 100
        description:
          title: Description
          type: string
          maxLength: 500
...
    ValidationError:
      title: ValidationError
      description: Schema for validation error item
      type: object
      required:
        - loc
        - msg
        - type
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string